fixing tailwind issue
